Moved NewsCard to components directory.
